Moving the slow query logging in a separate class
