+++ tests/TestCase/Database/Log/QueryLoggerTest.php
@@ -175,64 +175,4 @@ class QueryLoggerTest extends TestCase
         $engine2->expects($this->never())->method('log');
         $logger->log($query);
     }
-
-    /**
-     * Tests that only queries that run for a specific time are logged
-     *
-     * @return void
-     */
-    public function testThreshold()
-    {
-        $logger = $this->getMockBuilder(QueryLogger::class)
-            ->setConstructorArgs([
-                ['threshold' => 500]
-            ])
-            ->setMethods(['_log'])
-            ->getMock();
-
-        $slowQuery = new LoggedQuery;
-        $slowQuery->took = 5000;
-
-        $fastQuery = new LoggedQuery;
-        $fastQuery->took = 100;
-
-        $logger->expects($this->exactly(1))->method('_log');
-
-        $logger->log($fastQuery);
-        $logger->log($slowQuery);
-    }
-
-    /**
-     * Tests the filter callback
-     *
-     * @return void
-     */
-    public function testCallback()
-    {
-        $filter = function ($query) {
-            if ($query->query !== 'cakephp') {
-                return false;
-            }
-
-            return $query;
-        };
-
-        $logger = $this->getMockBuilder(QueryLogger::class)
-            ->setConstructorArgs([
-                ['filter' => $filter]
-            ])
-            ->setMethods(['_log'])
-            ->getMock();
-
-        $queryOne = new LoggedQuery;
-        $queryOne->query = 'cakephp';
-
-        $queryTwo = new LoggedQuery;
-        $queryTwo->query = 'something-else';
-
-        $logger->expects($this->exactly(1))->method('_log');
-
-        $logger->log($queryOne);
-        $logger->log($queryTwo);
-    }
 }
